Make sure to download all changes
By default gerrit limits the number of changes to 500 maximum. If a user
has more than 500 changes in the specified time range, then the changes
in the query are capped to 500, faking the actual stats. Therefore, to
get the real information, we need to tell gerrit to not limit the number
of changes in the query.